ESP-IDF: fix the 29 remaining example/test build failures
All 225 ESP-IDF apps (171 examples + 54 tests) now build on v6.1.

Drivers (shared by every C++ platform):
- BMP085/BMP180: the "0xFFFF = no device" calibration check compared
  the int16_t coefficients (AC1-3, B1, B2, MB, MC, MD) with 0xFFFF,
  which is never true, so an all-ones read went undetected for them.
  Compare as uint16_t.
- PCF8574/PCF8575: initializer list order now matches member
  declaration order (-Wreorder).

Examples:
- L3GD20H (all 3): used a nonexistent I2CConnectionESPIDF(port, addr)
  constructor; now sets up the bus like every other ESP-IDF I2C app.
  complete no longer calls who_am_i(), which isn't in the API (the
  constructor checks WHO_AM_I).
- BME680 complete/demo: FILTER_OFF -> FILTER_0.
- BMP280 complete: assign `alt` before using it.
- PCF8574 minimal: write_port() takes (port, mask).
- MCP2515 demo: cast the CAN id for %X (uint32_t is long on Xtensa).

Tests:
- l3gd20h: same bus setup fix; rewritten to the standard
  PASS/FAIL + ===DONE=== format so test_espidf.sh can parse it.
- pcf8591, l3g4200d, mpu9250, apds9960: removed checks that could
  never fail (e.g. uint8_t <= 255) and tripped -Wtype-limits.
- bmp581: dropped the check that read protected members; fixed
  BMP581Full.MODE_NORMAL -> BMP581Full::MODE_NORMAL.

+++ cpp/src/chips/pressure/BMP085.cpp
@@ -41,17 +41,17 @@ void BMP085Minimal::_read_calibration() {
     _mc  = (int16_t)((buf[18] << 8) | buf[19]);
     _md  = (int16_t)((buf[20] << 8) | buf[21]);
 
-    if (_ac1 == 0 || _ac1 == 0xFFFF ||
-        _ac2 == 0 || _ac2 == 0xFFFF ||
-        _ac3 == 0 || _ac3 == 0xFFFF ||
+    if (_ac1 == 0 || (uint16_t)_ac1 == 0xFFFF ||
+        _ac2 == 0 || (uint16_t)_ac2 == 0xFFFF ||
+        _ac3 == 0 || (uint16_t)_ac3 == 0xFFFF ||
         _ac4 == 0 || _ac4 == 0xFFFF ||
         _ac5 == 0 || _ac5 == 0xFFFF ||
         _ac6 == 0 || _ac6 == 0xFFFF ||
-        _b1  == 0 || _b1  == 0xFFFF ||
-        _b2  == 0 || _b2  == 0xFFFF ||
-        _mb  == 0 || _mb  == 0xFFFF ||
-        _mc  == 0 || _mc  == 0xFFFF ||
-        _md  == 0 || _md  == 0xFFFF) {
+        _b1  == 0 || (uint16_t)_b1 == 0xFFFF ||
+        _b2  == 0 || (uint16_t)_b2 == 0xFFFF ||
+        _mb  == 0 || (uint16_t)_mb == 0xFFFF ||
+        _mc  == 0 || (uint16_t)_mc == 0xFFFF ||
+        _md  == 0 || (uint16_t)_md == 0xFFFF) {
         abort();
     }
 }
